Move hooks and utils out of components
