+++ packages/ui/spa/utils/mergePatches.ts
